requirements.txt
requests
PyYAML
google-generativeai
